+++ app/src/main/java/com/github/fribourgsdp/radio/GameView.kt
@@ -56,7 +56,7 @@ interface GameView {
     /**
      * Display on the view the scores in the map [playerScores].
      */
-    fun displayPlayerScores(playerScores: Map<String, Int>)
+    fun displayPlayerScores(playerScores: Map<String, Long>)
 
     /**
      * An interface creating listeners able to handle songs pick in a [GameView].
