+++ app/src/main/java/com/github/fribourgsdp/radio/ScoresAdapter.kt
@@ -7,7 +7,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 
 class ScoresAdapter: RecyclerView.Adapter<ScoresAdapter.ViewHolder>() {
-    private var scoresList: List<Pair<String, Int>> = ArrayList()
+    private var scoresList: List<Pair<String, Long>> = ArrayList()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val itemView = LayoutInflater.from(parent.context).inflate(R.layout.card_score, parent, false)
@@ -25,7 +25,7 @@ class ScoresAdapter: RecyclerView.Adapter<ScoresAdapter.ViewHolder>() {
         return scoresList.size
     }
 
-    fun updateScore(scores: Map<String, Int>) {
+    fun updateScore(scores: Map<String, Long>) {
         scoresList = scores.toList().sortedBy { (_, score) -> score}
 
         // Update the whole list
